More work on 64-bit image support. Apparently I had not finished last time.
I made a pragmatic to prioritize 64-bit support over secure images. We had
merged in support for encrypted and signed images as V4 of the protocol.
But we never started using that version and it requires a fair amount of
infrastructure to support correctly. I did not want to have to build up that
infrastructure just to enable 64-bit images in the following version. So,
I vaporized V4 of the protocol and made 64-bit support be V5 which is now
the follow on to V3. Secure support will be readded as V6 or something after
that. The code is all still in there.

Back to 64-bit support. The basic tools: imagezip, imageunzip, imagedump,
and imagehash all work. The newer tools, that are based on the "libndz"
rewrite of the core code, do not. How is that for odd? V5 does include the
code for an optional per-image UUID that is put in every chunk. Since that
doesn't require any crypto libraries, it was easy to do. There are options
to imagezip/unzip to add/check a UUID.

Since imagehash required a new version as well, I made the hash field a bit
bigger and it now supports SHA256. Recall that imagehash is the tool that
breaks the image into 128K chunks and computes a hash of those chunks. It is
used for delta images, which means we don't use it right now.

This all compiles cleanly on FreeBSD (clang) right now, we will see how it
does with Linux. I seem to have issues with long vs. unsigned vs. long long
formats in printf. It will almost certainly hork if we tried to build on a
32-bit OS.

Things I still have not addressed:

* Hash code in imagezip. This is needed for creating delta images. Now that
  I have fixed the standalone hash program, this should be easy.
* Fix libndz and the related utilities.
* Replace the historic bubble sort used for ordering address ranges. Not
  directly related to 64-bitness, but necessary as images (and thus the number
  of ranges) gets bigger.
* More testing. So far I have only made sure it is backward compatible with
  32-bit images, I have not actually created one that needs 64-bit block
  numbers!

+++ imagezip/extfs/extfs.c
@@ -123,7 +123,8 @@ read_linuxslice(int slice, iz_type stype, iz_lba start, iz_size size,
 	if (excludenonfs && fssect < size) {
 		warnx("Linux Slice %d: filesystem smaller than partition, "
 		      "excluding [%lu-%lu]",
-		      dosslice, start+fssect, start+size-1);
+		      dosslice, (unsigned long)start+fssect,
+		      (unsigned long)start+size-1);
 		addskip(start + fssect, size - fssect);
 	}
 
@@ -366,8 +367,8 @@ read_linuxswap(int slice, iz_type stype, iz_lba start, iz_size size,
 		fprintf(stderr,
 			"  P%d (Linux Swap)\n", slice + 1 /* DOS Numbering */);
 		fprintf(stderr,
-			"        start %12ld, size %9ld\n",
-			start, size);
+			"        start %12lu, size %9lu\n",
+			(unsigned long)start, (unsigned long)size);
 	}
 
 	start += bytestosec(0x8000);
